Update design of room type settings
