+++ resources/js/components/RoomTypeSettingsField.vue
@@ -20,8 +20,14 @@ defineProps({
     </span>
 
     <div class="col-4 justify-content-center flex align-items-center">
-      <Tag v-if="type === 'switch' && value" icon="fas fa-check" rounded severity="primary" :aria-label="$t('app.enabled')"></Tag>
-      <Tag v-if="type === 'switch' && !value" icon="fas fa-times" rounded severity="secondary" :aria-label="$t('app.disabled')"></Tag>
+      <Tag v-if="type === 'switch' && value" class="round-tag" rounded severity="primary">
+        <span class="fas fa-check"  aria-hidden="true"></span>
+        <span class="sr-only">{{ $t('app.enabled') }}</span>
+      </Tag>
+      <Tag v-if="type === 'switch' && !value" class="round-tag" rounded severity="secondary">
+        <span class="fas fa-times" aria-hidden="true"></span>
+        <span class="sr-only">{{ $t('app.disabled') }}</span>
+      </Tag>
 
       <Tag severity="info" v-if="type === 'select'">
         {{ options[value]}}
@@ -29,3 +35,13 @@ defineProps({
     </div>
   </div>
 </template>
+<style scoped>
+.round-tag {
+  width: 1.5rem;
+  height: 1.5rem;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+</style>
