+++ resources/sass/app/app.scss
@@ -8,3 +8,4 @@
 @import "color-select";
 @import "streched-link";
 @import "menu";
+@import "accessibility";
